feature: Allow UglifyJS instance to be injected [fixes #77]
It's desirable to allow end users to inject their own version of
UglifyJS into the plugin code for two reasons:

1. Allow end users to update to newer versions of UglifyJS without
   waiting on this project to update, allowing them to easily test
   unreleased patches in the upstream project.
2. Allows us to more directly test our module's interface between gulp
   and UglifyJS, and to simplify our unit tests, since we no longer need
   to worry about the specific UglifyJS output.

This commit changes the implementation, allowing UglifyJS to be
injected, but does not change the tests to utilize this new feature.
This ensures we continue to pass our previous API contract.