feat(minifier): use UglifyJS 2.7.0's input map support
UglifyJS 2.7.0 adds support for a filename-content map as input, allowing
for the first time the ability to provide a filename for string input.

This change allows gulp-uglify to use UglifyJS's internal support for
source maps, which hopefully allows for better source maps for the entire
ecosystem.

BREAKING CHANGE: Now requires UglifyJS 2.7.0 or greater to use the
input object support.