fix(errors): emit errors of type `GulpUglifyError`
Use a lightweight wrapper to construct custom Error subtypes to replace
`PluginError`. Original errors from UglifyJS are now preserved, and
available as the "cause" property.

BREAKING CHANGE: Emitted errors are of a new type. Original UglifyJS
error message are preserved instead of the attempt to normalize them.
The constructor is exported as "GulpUglifyError" to allow for instance
checks.