fix(composer): invoke callback only once
In Node.js Streams, if an error is omitted downstream, that error can be
propagated back through connected streams. In gulp-uglify, this error
was being thrown by the next callback, being caught, and callback called
a second time.

This CL ensures the callback is called only once, by calling it outside
the try..catch block wrapping the minifier.

[terinjokes@gmail.com: amended change, rewrote tests and commit message]
Signed-off-by: Terin Stock <terinjokes@gmail.com>